tauri: refactor: separate main window webxdc commands
Move the webxdc commands that are supposed to be used
only by the main window (and not the webxdc apps themselves)
to a separate file.

This only moves code around, and does not make any changes
to the bodies of the functions.

This should make the separation of priviliges / permissions
easier to see and harder to mess up.
