+++ packages/target-tauri/src-tauri/src/lib.rs
@@ -128,13 +128,13 @@ pub fn run() {
             temp_file::write_temp_file,
             temp_file::remove_temp_file,
             temp_file::copy_blob_file_to_internal_tmp_dir,
-            webxdc::commands::on_webxdc_message_changed,
-            webxdc::commands::on_webxdc_message_deleted,
-            webxdc::commands::on_webxdc_status_update,
-            webxdc::commands::on_webxdc_realtime_data,
-            webxdc::commands::delete_webxdc_account_data,
-            webxdc::commands::close_all_webxdc_instances,
-            webxdc::commands::open_webxdc,
+            webxdc::commands_main_window::on_webxdc_message_changed,
+            webxdc::commands_main_window::on_webxdc_message_deleted,
+            webxdc::commands_main_window::on_webxdc_status_update,
+            webxdc::commands_main_window::on_webxdc_realtime_data,
+            webxdc::commands_main_window::delete_webxdc_account_data,
+            webxdc::commands_main_window::close_all_webxdc_instances,
+            webxdc::commands_main_window::open_webxdc,
             webxdc::commands::send_webxdc_update,
             webxdc::commands::get_webxdc_updates,
             webxdc::commands::join_webxdc_realtime_channel,
