+++ packages/target-tauri/src-tauri/src/webxdc/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod commands;
+pub(crate) mod commands_main_window;
 pub(crate) mod data_storage;
 mod error;
 pub(crate) mod icon_scheme;
